moved serializer/deserializer to detail dir/namespace
