+++ include/fkYAML/detail/deserializer.hpp
@@ -19,8 +19,8 @@
 #include "fkYAML/detail/version_macros.hpp"
 #include "fkYAML/detail/lexical_analyzer.hpp"
 #include "fkYAML/detail/type_traits.hpp"
+#include "fkYAML/detail/yaml_version_t.hpp"
 #include "fkYAML/exception.hpp"
-#include "fkYAML/node.hpp"
 
 /**
  * @namespace fkyaml
@@ -29,16 +29,22 @@
 FK_YAML_NAMESPACE_BEGIN
 
 /**
+ * @namespace detail
+ * @brief namespace for internal implementations of fkYAML library.
+ */
+namespace detail
+{
+
+/**
  * @class basic_deserializer
  * @brief A class which provides the feature of deserializing YAML documents.
  *
  * @tparam BasicNodeType A type of the container for deserialized YAML values.
  */
-template <typename BasicNodeType = node>
+template <typename BasicNodeType>
 class basic_deserializer
 {
-    static_assert(
-        detail::is_basic_node<BasicNodeType>::value, "basic_deserializer only accepts (const) basic_node<...>");
+    static_assert(detail::is_basic_node<BasicNodeType>::value, "basic_deserializer only accepts basic_node<...>");
 
     /** A type for sequence node value containers. */
     using sequence_type = typename BasicNodeType::sequence_type;
@@ -54,9 +60,6 @@ class basic_deserializer
     using string_type = typename BasicNodeType::string_type;
     /** A type for the lexical analyzer object used by this deserializer. */
     using lexer_type = detail::lexical_analyzer<BasicNodeType>;
-
-    using lexical_token_t = detail::lexical_token_t;
-    using yaml_version_t = detail::yaml_version_t;
 
 public:
     /**
@@ -321,10 +324,7 @@ private:
     std::unordered_map<std::string, BasicNodeType> m_anchor_table; /** The table of YAML anchor nodes. */
 };
 
-/**
- * @brief default YAML document deserializer.
- */
-using deserializer = basic_deserializer<>;
+} // namespace detail
 
 FK_YAML_NAMESPACE_END
